Vehicle filter
Add a new map filter control that opens a filter overlay, where the user
is able to pick which individual routes are going to be displayed.

If any subset is picked, a toggle radio button appears next to the
filter control. This allows the user to define the vehicle subset once
and either enable or disable the filter.

The filter is stored in local storage.

As a precaution measure, if there are no visible vehicles due to the
filter, the filter button will blink.